move nextgen to sep module
